Applied agent-q's patch to the SAGA pathfinding code for all platforms - x and y should not ever be greater than 640 and 480 respectively, so it looks safe enough to be applied
svn-id: r43500

+++ engines/saga/actor.h
@@ -183,8 +183,8 @@ enum DragonMoveTypes {
 
 struct PathDirectionData {
 	int8 direction;
-	int x;
-	int y;
+	int16 x;
+	int16 y;
 };
 
 struct ActorFrameRange {
